modules/dropai_python/python/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.6.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
dataclasses-json==0.6.7
deepdiff==7.0.1
emoji==2.12.1
exceptiongroup==1.2.1
filetype==1.2.0
Flask==2.2.5
flatbuffers==24.3.25
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-text-splitters==0.2.1
langdetect==1.0.9
langsmith==0.1.75
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.3
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
onnxruntime==1.18.0
opencv-python==4.10.0.82
opencv-python-headless==4.10.0.82
ordered-set==4.1.0
orjson==3.10.4
packaging==23.2
pillow==10.3.0
protobuf==5.27.1
pyclipper==1.3.0.post5
pydantic==2.7.3
pydantic_core==2.18.4
PyMuPDF==1.24.5
PyMuPDFb==1.24.3
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-iso639==2024.4.27
python-magic==0.4.27
python-pptx==0.6.23
PyYAML==6.0.1
rapidfuzz==3.9.3
rapidocr-onnxruntime==1.3.22
rapidocr-pdf==0.1.0
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
semantic-text-splitter==0.13.3
shapely==2.0.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sympy==1.12.1
tabulate==0.9.0
tenacity==8.3.0
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
unstructured==0.14.5
unstructured-client==0.23.3
urllib3==2.2.1
Werkzeug==2.2.3
wrapt==1.16.0
XlsxWriter==3.2.0
yarl==1.9.4
zipp==3.15.0
